fix: support content page toggling
