+++ sites/public/src/pages/privacy.tsx
@@ -1,31 +1,7 @@
-import React, { useEffect, useContext } from "react"
-import { t } from "@bloom-housing/ui-components"
-import Markdown from "markdown-to-jsx"
-import { PageView, pushGtmEvent, AuthContext } from "@bloom-housing/shared-helpers"
-import { UserStatus } from "../lib/constants"
-import Layout from "../layouts/application"
-import pageContent from "../md_content/privacy_policy.md"
-import { PageHeaderLayout } from "../patterns/PageHeaderLayout"
-import styles from "../patterns/PageHeaderLayout.module.scss"
+import React from "react"
+import PrivacyDeprecated from "../components/content-pages/PrivacyDeprecated"
+import PrivacySeeds from "../components/content-pages/PrivacySeeds"
 
-const Privacy = () => {
-  const { profile } = useContext(AuthContext)
-
-  useEffect(() => {
-    pushGtmEvent<PageView>({
-      event: "pageView",
-      pageTitle: "Privacy",
-      status: profile ? UserStatus.LoggedIn : UserStatus.NotLoggedIn,
-    })
-  }, [profile])
-
-  return (
-    <Layout>
-      <PageHeaderLayout heading={t("pageTitle.privacy")} inverse>
-        <Markdown className={styles["markdown"]}>{pageContent.toString()}</Markdown>
-      </PageHeaderLayout>
-    </Layout>
-  )
-}
+const Privacy = () => (process.env.showNewSeedsDesigns ? <PrivacySeeds /> : <PrivacyDeprecated />)
 
 export default Privacy
